requirements.txt
mcp
mcp[cli]
mcp-use
azure-cosmos
python-dotenv
openai
nltk
streamlit
langchain-openai